full working converter
to get working; you will need to install the dependencies in package.json, not sure if you need to install another db but if the db.ts is throwing errors you'll need to bash: npm i --save-dev @types/better-sqlite3.
you will also need to create a .env file in the main project folder with the text i emailed to you.